#20-그룹 아이디 수정 및 패키지 구조 수정
그룹아이디를 com.sparta에서 com.sparta.newsfeed로 수정 및 도메인 별로 패키지 만들어관리하고자 전체 구조 수정
