--- src/main/java/com/sparta/NewsFeedApplication.java
@@ -1,15 +1,0 @@
-package com.sparta;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class NewsFeedApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(NewsFeedApplication.class, args);
-    }
-
-}
