--- src/main/java/com/sparta/config/PasswordConfig.java
@@ -1,16 +1,0 @@
-package com.sparta.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class PasswordConfig {
-
-    @Bean
-    public PasswordEncoder passwordEncoder(){
-        // BCrytPassword사용 메서드
-        return new BCryptPasswordEncoder();
-    }
-}
