--- src/main/java/com/sparta/domain/BaseEntity.java
@@ -1,25 +1,0 @@
-package com.sparta.domain;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDate;
-
-@MappedSuperclass
-@EntityListeners(value = {AuditingEntityListener.class})
-@Getter
-abstract class BaseEntity {
-    @CreatedDate
-    @Column(name="created_at", updatable = false)
-    private LocalDate createdAt;
-
-    @LastModifiedDate
-    @Column(name="modified_at")
-    private LocalDate modifiedAt;
-}
