--- src/main/java/com/sparta/domain/dto/RequestProfileUpdateDto.java
@@ -1,12 +1,0 @@
-package com.sparta.domain.dto;
-
-import lombok.Getter;
-
-@Getter
-public class RequestProfileUpdateDto {
-
-    private String username;
-    private String password;
-    private String content;
-
-}
