--- src/main/java/com/sparta/domain/dto/SignupDto.java
@@ -1,10 +1,0 @@
-package com.sparta.domain.dto;
-
-import lombok.Getter;
-
-@Getter
-public class SignupDto {
-    private String username;
-    private String password;
-
-}
