--- src/main/java/com/sparta/dto/CommentRequestDto.java
@@ -1,8 +1,0 @@
-package com.sparta.dto;
-
-import lombok.Getter;
-
-@Getter
-public class CommentRequestDto {
-    String content;
-}
