+++ src/main/java/com/sparta/newsfeed/NewsFeedApplication.java
@@ -1,0 +1,15 @@
+package com.sparta.newsfeed;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableJpaAuditing
+public class NewsFeedApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(NewsFeedApplication.class, args);
+    }
+
+}
