+++ src/main/java/com/sparta/newsfeed/board/repository/BoardRepository.java
@@ -1,0 +1,10 @@
+package com.sparta.newsfeed.board.repository;
+
+import com.sparta.newsfeed.board.domain.Board;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface BoardRepository extends JpaRepository<Board, Long> {
+    List<Board> findAllByOrderByIdDesc();
+}
