+++ src/main/java/com/sparta/newsfeed/comment/dto/CommentRequestDto.java
@@ -1,0 +1,12 @@
+package com.sparta.newsfeed.comment.dto;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+
+
+@Data
+@Builder
+public class CommentRequestDto {
+    String content;
+}
