+++ src/main/java/com/sparta/newsfeed/domain/BaseEntity.java
@@ -1,0 +1,26 @@
+package com.sparta.newsfeed.domain;
+
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+
+@MappedSuperclass
+@EntityListeners(value = {AuditingEntityListener.class})
+@Getter
+public abstract class BaseEntity {
+    @CreatedDate
+    @Column(name="created_at", updatable = false)
+    private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    @Column(name="modified_at")
+    private LocalDateTime modifiedAt;
+}
