+++ src/main/java/com/sparta/newsfeed/member/dto/RequestProfileUpdateDto.java
@@ -1,0 +1,12 @@
+package com.sparta.newsfeed.member.dto;
+
+import lombok.Getter;
+
+@Getter
+public class RequestProfileUpdateDto {
+
+    private String username;
+    private String password;
+    private String content;
+
+}
