+++ src/main/java/com/sparta/newsfeed/member/dto/SignupDto.java
@@ -1,0 +1,11 @@
+package com.sparta.newsfeed.member.dto;
+
+import lombok.Getter;
+
+@Getter
+public class SignupDto {
+    private String username;
+    private String password;
+    private String content;
+
+}
