+++ src/main/java/com/sparta/newsfeed/security/config/PasswordConfig.java
@@ -1,0 +1,16 @@
+package com.sparta.newsfeed.security.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class PasswordConfig {
+
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        // BCrytPassword사용 메서드
+        return new BCryptPasswordEncoder();
+    }
+}
