+++ src/main/java/com/sparta/newsfeed/security/service/MemberDetailsImpl.java
@@ -1,0 +1,62 @@
+package com.sparta.newsfeed.security.service;
+
+import com.sparta.newsfeed.member.domain.Member;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+public class MemberDetailsImpl implements UserDetails {
+
+    private final Member member;
+
+    public MemberDetailsImpl(Member member) {
+        this.member = member;
+    }
+
+    public Member getMember() {
+        return member;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        /*UserRoleEnum role = user.getRole();
+        String authority = role.getAuthority();
+
+        SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(authority);
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(simpleGrantedAuthority);
+        */
+        return null;
+    }
+
+    @Override
+    public String getPassword() {
+        return member.getPassword();
+    }
+    // 오류 지울려고 getter, setter Member에 설정
+    @Override
+    public String getUsername() {
+        return member.getUsername();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
