+++ src/main/java/com/sparta/newsfeed/security/service/MemberDetailsServiceImpl.java
@@ -1,0 +1,27 @@
+package com.sparta.newsfeed.security.service;
+
+import com.sparta.newsfeed.member.domain.Member;
+import com.sparta.newsfeed.member.repository.MemberRepository;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class MemberDetailsServiceImpl implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    public MemberDetailsServiceImpl(MemberRepository memberRepository) {
+        this.memberRepository = memberRepository;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Member member = memberRepository.findById(username).orElseThrow(
+                ()-> new UsernameNotFoundException("Not Found " + username)
+        );
+
+        return new MemberDetailsImpl(member);
+    }
+}
