--- src/main/java/com/sparta/repository/BoardRepository.java
@@ -1,10 +1,0 @@
-package com.sparta.repository;
-
-import com.sparta.domain.Board;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface BoardRepository extends JpaRepository<Board, Long> {
-    List<Board> findAllByOrderByIdDesc();
-}
