--- src/main/java/com/sparta/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.sparta.repository;
-
-import com.sparta.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, String> {
-    //Insert테스트를 위한 메소드(임시)
-    Long countBy();
-}
