--- src/main/java/com/sparta/security/MemberDetailsServiceImpl.java
@@ -1,27 +1,0 @@
-package com.sparta.security;
-
-import com.sparta.domain.Member;
-import com.sparta.repository.MemberRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MemberDetailsServiceImpl implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-
-    public MemberDetailsServiceImpl(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Member member = memberRepository.findByUsername(username).orElseThrow(
-                ()-> new UsernameNotFoundException("Not Found " + username)
-        );
-
-        return new MemberDetailsImpl(member);
-    }
-}
