+++ src/test/java/com/sparta/newsfeed/NewsFeedApplicationTests.java
@@ -1,0 +1,14 @@
+package com.sparta.newsfeed;
+
+
+import lombok.RequiredArgsConstructor;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@RequiredArgsConstructor
+@SpringBootTest
+class NewsFeedApplicationTests {
+    @Test
+    void contextLoads() {
+    }
+}
