--- src/test/java/com/sparta/repository/CommentRepositoryTest.java
@@ -1,17 +1,0 @@
-package com.sparta.repository;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@SpringBootTest
-class CommentRepositoryTest {
-    @Autowired
-    CommentRepository commentRepository;
-
-    @Test
-    void name() {
-    }
-}
